refactor: #117 Dashboard サービス層を Core/Observability へ移動
- ConfigurationService / IConfigurationService / ConfigDto / ConfigUpdateDto
  → CloudMigrator.Core.Configuration
- SetupDoctorService / ISetupDoctorService / DoctorOptions / DoctorResult / DoctorCheck
  → CloudMigrator.Core.Setup
- ILogChannel / LogChannelAdapter
  → CloudMigrator.Observability
- TransferJobService / ITransferJobService / TransferJobInfo / JobStatus
  → CloudMigrator.Core.Transfer

- CloudMigrator.Core.csproj: Microsoft.Extensions.Http を追加（IHttpClientFactory）
- Dashboard/_Imports.razor: 新 namespace の @using を追加
- Dashboard/App.xaml.cs: 新 namespace の using を追加
- codecov.yml: ignore を Dashboard UI コードのみに絞り込み
- tests/unit: ConfigurationServiceTests / SetupDoctorServiceTests / TransferJobServiceTests の using を更新
