+++ src/CloudMigrator.Core/Setup/SetupDoctorService.cs
@@ -3,7 +3,7 @@ using System.Net.Http.Headers;
 using System.Text.Json;
 using Microsoft.Extensions.Logging;
 
-namespace CloudMigrator.Dashboard;
+namespace CloudMigrator.Core.Setup;
 
 /// <summary>
 /// セットアップ診断サービスのオプション。
