+++ src/CloudMigrator.Dashboard/_Imports.razor
@@ -7,7 +7,10 @@
 @using Microsoft.AspNetCore.Components.Web
 @using Microsoft.Extensions.DependencyInjection
 @using MudBlazor
+@using CloudMigrator.Core.Configuration
+@using CloudMigrator.Core.Setup
 @using CloudMigrator.Core.State
+@using CloudMigrator.Core.Transfer
 @using CloudMigrator.Observability
 @using CloudMigrator.Dashboard
 @using CloudMigrator.Dashboard.Components
